Extract commonality from (Java|Groovy)LibraryDescription
Summary:

These two classes are in fact more similar than their implementations
suggest. In this change we extract the common features to both Groovy
and Java into JvmLibraryDescription, and reduce the previous classes
to smaller configuration classes. This removes a lot of duplication
while also adding flavours and maven support to groovy modules
(documentation to follow).

Test-plan:

CI
