+++ src/com/facebook/buck/jvm/groovy/GroovyLibraryDescription.java
@@ -16,39 +16,29 @@
 
 package com.facebook.buck.jvm.groovy;
 
-import static com.facebook.buck.jvm.common.ResourceValidator.validateResources;
-
-import com.facebook.buck.jvm.java.CalculateAbi;
-import com.facebook.buck.jvm.java.DefaultJavaLibrary;
 import com.facebook.buck.jvm.java.JavacOptions;
 import com.facebook.buck.jvm.java.JavacOptionsFactory;
-import com.facebook.buck.jvm.java.JvmLibraryArg;
-import com.facebook.buck.model.BuildTarget;
+import com.facebook.buck.jvm.java.JvmLibraryConfigurable;
+import com.facebook.buck.jvm.java.JvmLibraryConfiguration;
+import com.facebook.buck.jvm.java.JvmLibraryDescription;
 import com.facebook.buck.rules.BuildRule;
 import com.facebook.buck.rules.BuildRuleParams;
 import com.facebook.buck.rules.BuildRuleResolver;
 import com.facebook.buck.rules.BuildRuleType;
-import com.facebook.buck.rules.BuildRules;
-import com.facebook.buck.rules.BuildTargetSourcePath;
-import com.facebook.buck.rules.Description;
-import com.facebook.buck.rules.SourcePath;
 import com.facebook.buck.rules.SourcePathResolver;
-import com.facebook.buck.rules.TargetGraph;
 import com.facebook.infer.annotation.SuppressFieldNotInitialized;
 import com.google.common.base.Optional;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.ImmutableSortedSet;
-import com.google.common.collect.Iterables;
 
 import java.nio.file.Path;
 
-public class GroovyLibraryDescription implements Description<GroovyLibraryDescription.Arg> {
+public class GroovyLibraryDescription
+    implements JvmLibraryConfigurable<GroovyLibraryDescription.Arg> {
 
   public static final BuildRuleType TYPE = BuildRuleType.of("groovy_library");
 
   private final GroovyBuckConfig groovyBuckConfig;
-  // For cross compilation
+  // For cross compilation.
   private final JavacOptions defaultJavacOptions;
 
   public GroovyLibraryDescription(
@@ -56,6 +46,29 @@ public class GroovyLibraryDescription implements Description<GroovyLibraryDescri
       JavacOptions defaultJavacOptions) {
     this.groovyBuckConfig = groovyBuckConfig;
     this.defaultJavacOptions = defaultJavacOptions;
+  }
+
+  @Override
+  public JvmLibraryConfiguration createConfiguration(
+      BuildRuleParams params,
+      BuildRuleResolver resolver,
+      SourcePathResolver pathResolver,
+      GroovyLibraryDescription.Arg args) {
+
+    GroovycToJarStepFactory groovycToJarStepFactory = new GroovycToJarStepFactory(
+        groovyBuckConfig.getGroovyCompiler().get(),
+        args.extraGroovycArguments,
+        JavacOptionsFactory.create(
+            defaultJavacOptions,
+            params,
+            resolver,
+            pathResolver,
+            args
+        ));
+    return new JvmLibraryConfiguration(
+        groovycToJarStepFactory,
+        Optional.<Path>absent(),
+        ImmutableList.<BuildRule>of());
   }
 
   @Override
@@ -68,72 +81,8 @@ public class GroovyLibraryDescription implements Description<GroovyLibraryDescri
     return new Arg();
   }
 
-  @Override
-  public <A extends Arg> BuildRule createBuildRule(
-      TargetGraph targetGraph,
-      BuildRuleParams params,
-      BuildRuleResolver resolver,
-      A args) {
-    SourcePathResolver pathResolver = new SourcePathResolver(resolver);
-
-    BuildTarget abiJarTarget =
-        BuildTarget.builder(params.getBuildTarget())
-            .addFlavors(CalculateAbi.FLAVOR)
-            .build();
-
-    ImmutableSortedSet<BuildRule> exportedDeps = resolver.getAllRules(args.exportedDeps.get());
-    DefaultJavaLibrary defaultJavaLibrary =
-        resolver.addToIndex(
-            new DefaultJavaLibrary(
-                params.appendExtraDeps(
-                        BuildRules.getExportedRules(
-                            Iterables.concat(
-                                params.getDeclaredDeps().get(),
-                                exportedDeps,
-                                resolver.getAllRules(args.providedDeps.get())))),
-                pathResolver,
-                args.srcs.get(),
-                validateResources(
-                    pathResolver,
-                    params.getProjectFilesystem(),
-                    args.resources.get()),
-                Optional.<Path>absent(),
-                Optional.<SourcePath>absent(),
-                ImmutableList.<String>of(),
-                exportedDeps,
-                resolver.getAllRules(args.providedDeps.get()),
-                new BuildTargetSourcePath(abiJarTarget),
-                /* additionalClasspathEntries */ ImmutableSet.<Path>of(),
-                new GroovycToJarStepFactory(
-                    groovyBuckConfig.getGroovyCompiler().get(),
-                    args.extraGroovycArguments,
-                    JavacOptionsFactory.create(
-                        defaultJavacOptions,
-                        params,
-                        resolver,
-                        pathResolver,
-                        args)),
-                Optional.<Path>absent(),
-                Optional.<String>absent(),
-                ImmutableSortedSet.<BuildTarget>of()));
-
-    resolver.addToIndex(
-        CalculateAbi.of(
-            abiJarTarget,
-            pathResolver,
-            params,
-            new BuildTargetSourcePath(defaultJavaLibrary.getBuildTarget())));
-
-    return defaultJavaLibrary;
-  }
-
   @SuppressFieldNotInitialized
-  public static class Arg extends JvmLibraryArg {
-    public Optional<ImmutableSortedSet<SourcePath>> srcs;
-    public Optional<ImmutableSortedSet<SourcePath>> resources;
+  public static class Arg extends JvmLibraryDescription.Arg {
     public Optional<ImmutableList<String>> extraGroovycArguments;
-    public Optional<ImmutableSortedSet<BuildTarget>> providedDeps;
-    public Optional<ImmutableSortedSet<BuildTarget>> exportedDeps;
-    public Optional<ImmutableSortedSet<BuildTarget>> deps;
   }
 }
