+++ src/com/facebook/buck/jvm/java/BUCK
@@ -139,6 +139,9 @@ java_library(
     'JavaSourceJar.java',
     'JavaTest.java',
     'JavaTestDescription.java',
+    'JvmLibraryConfigurable.java',
+    'JvmLibraryConfiguration.java',
+    'JvmLibraryDescription.java',
     'Keystore.java',
     'KeystoreDescription.java',
     'MavenUberJar.java',
