+++ src/com/facebook/buck/jvm/java/JvmLibraryConfiguration.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2015-present Facebook, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License. You may obtain
+ * a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+
+package com.facebook.buck.jvm.java;
+
+import com.facebook.buck.rules.BuildRule;
+import com.google.common.base.Optional;
+
+import java.nio.file.Path;
+import java.util.Collection;
+
+public class JvmLibraryConfiguration {
+  public final CompileToJarStepFactory compileToJarStepFactory;
+  public final Optional<Path> generatedSourceDirectory;
+  public final Collection<BuildRule> additionalBuildRules;
+
+  public JvmLibraryConfiguration(
+      CompileToJarStepFactory compileToJarStepFactory,
+      Optional<Path> generatedSourceDirectory,
+      Collection<BuildRule> additionalBuildRules) {
+    this.compileToJarStepFactory = compileToJarStepFactory;
+    this.generatedSourceDirectory = generatedSourceDirectory;
+    this.additionalBuildRules = additionalBuildRules;
+  }
+}
